pyproject.toml
[project]
name = "ctf-agent-two"
version = "0.1.0"
description = "CTF agent v2 — multi-model solver swarm"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai[bedrock,openai,google]",
    "pydantic-settings",
    "genai-prices",
    "boto3",
    "aiodocker",
    "claude-agent-sdk",
    "httpx",
    "pyyaml",
    "rich",
    "click",
    "markdownify",
]

[project.scripts]
ctf-solve = "backend.cli:main"
ctf-msg = "backend.cli:msg"
ctf-pull = "backend.pull_challenges_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "ruff>=0.14",
    "ty>=0.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N", "C4", "PTH"]
ignore = ["E501", "SIM105", "SIM115", "PTH123"]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.ty.rules]
unresolved-import = "ignore"

[tool.ty.environment]
python-version = "3.12"
